Share analysis node JSON serialization
